Reorganise scroll utilities
This splits the scroll utilities into more files, in an effort to keep things organised and make space for further scrolling-related utilities.
